Fix build
git-svn-id: http://svn.code.sf.net/p/htmlunit/code/trunk/htmlunit@13376 5f5364db-9458-4db8-a492-e30667be6df6

+++ src/main/java/com/gargoylesoftware/htmlunit/html/HtmlNumberInput.java
@@ -192,7 +192,7 @@ public class HtmlNumberInput extends HtmlInput implements SelectableTextInput {
     @Override
     public void setValueAttribute(final String newValue) {
         try {
-            Integer.parseInt(newValue);
+            Long.parseLong(newValue);
             super.setValueAttribute(newValue);
         }
         catch (final NumberFormatException e) {
